Changed all the doubles to decimal

+++ ZasUndDas.Shared/Data/CalzoneDTO.cs
@@ -8,6 +8,6 @@ public class CalzoneDTO
     public int Id { set; get; }
     public int? SauceId { set; get; }
     public bool? CookedAtHome { set; get; }
-    public double Price { set; get; }
+    public decimal Price { set; get; }
     public virtual Sauce? Sauce { set; get; }
 }
